IGNITE-5729 - IgniteCacheProxy instances from with() methods are not reusable after cache restart
